feat(frontend): rename all dashboard list page
Looking for a single 's' to determine singular or plural dashboard page was a bother
